Added custom fields conversion tool `hf-dewpcf`.
The original session metadata was entered as wordpress metadata, using
a plugin called `wordpress custom fields`.

This created two issues with maintaining the site.

1. The metadata keys were obtuse and prefixed.
2. Dates were stored as timestamps.

This tool converts the target keys and timestamp data to a new format,
and removes the old metadata.

It was run once against a transition theme, that worked with both key
formats.